+++ examples/ExpoMessaging/App.js
@@ -1,5 +1,6 @@
 import React, { PureComponent } from 'react';
 import { View, SafeAreaView, TouchableOpacity, Text } from 'react-native';
+
 import { StreamChat } from 'stream-chat';
 import {
   Chat,
@@ -10,19 +11,24 @@ import {
   Thread,
   CloseButton,
   ChannelPreviewMessenger,
+  Streami18n
 } from 'stream-chat-expo';
 
 import { createAppContainer, createStackNavigator } from 'react-navigation';
 
-const chatClient = new StreamChat('qk4nn7rpcn75');
+const chatClient = new StreamChat('q95x9hkbyd6p');
 const userToken =
-  'eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJ1c2VyX2lkIjoiYmlsbG93aW5nLWZpcmVmbHktOCJ9.CQTVyJ6INIM8u28BxkneY2gdYpamjLzSVUOTZKzfQlg';
-const user = {
-  id: 'billowing-firefly-8',
-  name: 'Billowing firefly',
-  image:
-    'https://stepupandlive.files.wordpress.com/2014/09/3d-animated-frog-image.jpg',
-};
+  'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJ1c2VyX2lkIjoicm9uIn0.eRVjxLvd4aqCEHY_JRa97g6k7WpHEhxL7Z4K4yTot1c';
+  const user = {
+    id: 'ron',
+  };
+  
+const filters = { type: 'messaging', example: 'example-apps', members: { '$in': ['ron'] } };
+const sort = { last_message_at: -1 };
+const options = {
+  state: true,
+  watch: true
+}
 // Read more about style customizations at - https://getstream.io/chat/react-native-chat/tutorial/#custom-styles
 const theme = {
   avatar: {
@@ -31,7 +37,7 @@ const theme = {
     },
   },
   colors: {
-    primary: 'magenta',
+    primary: 'blue',
   },
   spinner: {
     css: `
@@ -41,14 +47,13 @@ const theme = {
   },
 };
 
-
-const filters = { type: 'messaging' };
-const sort = { last_message_at: -1 };
-const options = {
-  state: true,
-  watch: true,
-  limit: 30,
-};
+/**
+ * Start playing with streami18n instance here:
+ * Please refer to description of this PR for details: https://github.com/GetStream/stream-chat-react-native/pull/150
+ */
+const streami18n = new Streami18n({
+  language: 'en'
+});
 
 class ChannelListScreen extends PureComponent {
   static navigationOptions = () => ({
@@ -58,7 +63,7 @@ class ChannelListScreen extends PureComponent {
   render() {
     return (
       <SafeAreaView>
-        <Chat client={chatClient} style={theme}>
+        <Chat client={chatClient} style={theme} i18nInstance={streami18n}>
           <View style={{ display: 'flex', height: '100%', padding: 10 }}>
             <ChannelList
               filters={filters}
@@ -94,7 +99,7 @@ class ChannelScreen extends PureComponent {
 
     return (
       <SafeAreaView>
-        <Chat client={chatClient} style={theme}>
+        <Chat client={chatClient} style={theme} i18nInstance={streami18n}>
           <Channel client={chatClient} channel={channel}>
             <View style={{ display: 'flex', height: '100%' }}>
               <MessageList
@@ -149,7 +154,7 @@ class ThreadScreen extends PureComponent {
 
     return (
       <SafeAreaView>
-        <Chat client={chatClient}>
+        <Chat client={chatClient} i18nInstance={streami18n}>
           <Channel
             client={chatClient}
             channel={channel}